+++ docs/changelog.rst
@@ -10,6 +10,7 @@ CHANGELOG
   See settings ``TREK_EXPORT_INFORMATION_DESK_LIST_LIMIT`` and ``TREK_EXPORT_POI_LIST_LIMIT``
 * Justify texts of POIs in trek export, now converted to plain text.
 * Trek export geometries are now translucid red by default (see ``MAP_STYLES`` setting) (ref #1120)
+* Paths apparence in trek exports are now controlled by MAP_STYLES setting too.
 
 
 0.25.2 (2014-08-11)